+++ TripoliApp/src/main/java/org/cirdles/tripoli/gui/TripoliGUI.java
@@ -32,7 +32,6 @@ import javafx.stage.Stage;
 import javafx.stage.Window;
 import javafx.stage.WindowEvent;
 
-import java.awt.*;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.PrintStream;
@@ -71,14 +70,13 @@ public class TripoliGUI extends Application {
 //    |_____| [___]   [___]| ;.__/  '.__.'[___][___]
 //                        [__|
 
-        StringBuilder logo = new StringBuilder();
-        logo.append("        _________          _                  __    _   \n");
-        logo.append("       |  _   _  |        (_)                [  |  (_)  \n");
-        logo.append("       |_/ | | \\_|_ .--.  __  _ .--.    .--.  | |  __   \n");
-        logo.append("           | |   [ `/'`\\][  |[ '/'`\\ \\/ .'`\\ \\| | [  |  \n");
-        logo.append("          _| |_   | |     | | | \\__/ || \\__. || |  | |  \n");
-        logo.append("         |_____| [___]   [___]| ;.__/  '.__.'[___][___] \n");
-        logo.append("                             [__|                       \n");
+        String logo = "        _________          _                  __    _   \n" +
+                "       |  _   _  |        (_)                [  |  (_)  \n" +
+                "       |_/ | | \\_|_ .--.  __  _ .--.    .--.  | |  __   \n" +
+                "           | |   [ `/'`\\][  |[ '/'`\\ \\/ .'`\\ \\| | [  |  \n" +
+                "          _| |_   | |     | | | \\__/ || \\__. || |  | |  \n" +
+                "         |_____| [___]   [___]| ;.__/  '.__.'[___][___] \n" +
+                "                             [__|                       \n";
         System.out.println(logo);
