+++ TripoliApp/src/main/java/org/cirdles/tripoli/gui/dataViews/plots/plotsControllers/tripoliPlots/BasicScatterPlot.java
@@ -1,50 +1,37 @@
-package org.cirdles.tripoli.gui.dataViews.plots;
+package org.cirdles.tripoli.gui.dataViews.plots.plotsControllers.tripoliPlots;
 
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Paint;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.text.Font;
 import javafx.scene.text.Text;
+import org.cirdles.tripoli.gui.dataViews.plots.AbstractDataView;
+import org.cirdles.tripoli.gui.dataViews.plots.TicGeneratorForAxes;
 import org.cirdles.tripoli.visualizationUtilities.linePlots.LinePlotBuilder;
 
-import java.math.BigDecimal;
-import java.math.RoundingMode;
-import java.util.ArrayList;
-import java.util.List;
 
+public class BasicScatterPlot extends AbstractDataView {
 
-public class BasicLinePlotLogX extends AbstractDataView {
-
-    private final LinePlotBuilder linePlotBuilder;
+    private final LinePlotBuilder intensityLinePlotBuilder;
 
     /**
      * @param bounds
-     * @param linePlotBuilder
+     * @param intensityLinePlotBuilder
      */
-    public BasicLinePlotLogX(Rectangle bounds, LinePlotBuilder linePlotBuilder) {
+    public BasicScatterPlot(Rectangle bounds, LinePlotBuilder intensityLinePlotBuilder) {
         super(bounds, 50, 5);
-        this.linePlotBuilder = linePlotBuilder;
+        this.intensityLinePlotBuilder = intensityLinePlotBuilder;
     }
 
     @Override
     public void preparePanel() {
-        xAxisData = linePlotBuilder.getxData();
-        yAxisData = linePlotBuilder.getyData();
+        xAxisData = intensityLinePlotBuilder.getxData();
+        yAxisData = intensityLinePlotBuilder.getyData();
 
-        minX = Math.log(xAxisData[0]);
-        maxX = Math.log(xAxisData[xAxisData.length - 1]);
+        minX = xAxisData[0];
+        maxX = xAxisData[xAxisData.length - 1];
 
-        // logarithmic ticsX
-        List<Double> xTicsList = new ArrayList<>();
-        int limitLog = (int) xAxisData[xAxisData.length - 1];
-        for (int logIndex = 1; logIndex <= limitLog; logIndex = logIndex * 10) {
-            xTicsList.add(Math.log(logIndex));
-        }
-        ticsX = new BigDecimal[xTicsList.size()];
-        for (int i = 0; i < xTicsList.size(); i++) {
-            ticsX[i] = new BigDecimal(Double.toString(xTicsList.get(i)));
-        }
-
+        ticsX = TicGeneratorForAxes.generateTics(minX, maxX, (int) (graphWidth / 40.0));
         double xMarginStretch = TicGeneratorForAxes.generateMarginAdjustment(minX, maxX, 0.01);
         minX -= xMarginStretch;
         maxX += xMarginStretch;
@@ -81,19 +68,15 @@ public class BasicLinePlotLogX extends AbstractDataView {
         text.setFont(Font.font("SansSerif", 12));
         int textWidth = 0;
 
-        labelXAxis("Log of Saved Iteration Count");
-        showTitle(linePlotBuilder.getTitle());
+        showTitle(intensityLinePlotBuilder.getTitle());
 
-        g2d.setLineWidth(2.0);
-        // new line plot
+        g2d.setLineWidth(0.5);
+        // scatter plot
         g2d.setStroke(Paint.valueOf("Black"));
-        g2d.beginPath();
-        g2d.moveTo(mapX(Math.log(xAxisData[0])), mapY(yAxisData[0]));
         for (int i = 0; i < xAxisData.length; i++) {
-            // line tracing through points
-            g2d.lineTo(mapX(Math.log(xAxisData[i])), mapY(yAxisData[i]));
+            g2d.strokeOval(mapX(xAxisData[i]) - 2f, mapY(yAxisData[i]) - 2f, 4f, 4f);
         }
-        g2d.stroke();
+
 
         if (ticsY.length > 1) {
             // border and fill
@@ -111,7 +94,7 @@ public class BasicLinePlotLogX extends AbstractDataView {
             float verticalTextShift = 3.2f;
             g2d.setFont(Font.font("SansSerif", 10));
             if (ticsY != null) {
-                for (BigDecimal bigDecimal : ticsY) {
+                for (java.math.BigDecimal bigDecimal : ticsY) {
                     g2d.strokeLine(
                             mapX(minX), mapY(bigDecimal.doubleValue()), mapX(maxX), mapY(bigDecimal.doubleValue()));
 
@@ -119,24 +102,24 @@ public class BasicLinePlotLogX extends AbstractDataView {
                     text.setText(bigDecimal.toString());
                     textWidth = (int) text.getLayoutBounds().getWidth();
                     g2d.fillText(text.getText(),//
-                            (float) mapX(minX) - textWidth - 5f,
+                            (float) mapX(minX) - textWidth + 5f,
                             (float) mapY(bigDecimal.doubleValue()) + verticalTextShift);
 
                 }
                 // ticsX
                 if (ticsX != null) {
-                    for (BigDecimal bigDecimal : ticsX) {
+                    for (int i = 0; i < ticsX.length - 1; i++) {
                         try {
                             g2d.strokeLine(
-                                    mapX(bigDecimal.doubleValue()),
+                                    mapX(ticsX[i].doubleValue()),
                                     mapY(ticsY[0].doubleValue()),
-                                    mapX(bigDecimal.doubleValue()),
+                                    mapX(ticsX[i].doubleValue()),
                                     mapY(ticsY[0].doubleValue()) + 5);
 
                             // bottom
-                            String xText = (new BigDecimal(Double.toString(Math.exp(bigDecimal.doubleValue())))).setScale(-1, RoundingMode.HALF_UP).toPlainString();
+                            String xText = ticsX[i].toPlainString();
                             g2d.fillText(xText,
-                                    (float) mapX(bigDecimal.doubleValue()) - 5f,
+                                    (float) mapX(ticsX[i].doubleValue()) - 5f,
                                     (float) mapY(ticsY[0].doubleValue()) + 15);
 
                         } catch (Exception ignored) {
