+++ TripoliApp/src/main/java/org/cirdles/tripoli/gui/dataViews/plots/plotsControllers/tripoliPlots/MultiLinePlotLogX.java
@@ -1,37 +1,53 @@
-package org.cirdles.tripoli.gui.dataViews.plots;
+package org.cirdles.tripoli.gui.dataViews.plots.plotsControllers.tripoliPlots;
 
 import javafx.scene.canvas.GraphicsContext;
+import javafx.scene.paint.Color;
 import javafx.scene.paint.Paint;
 import javafx.scene.shape.Rectangle;
 import javafx.scene.text.Font;
 import javafx.scene.text.Text;
-import org.cirdles.tripoli.visualizationUtilities.linePlots.ComboPlotBuilder;
+import org.cirdles.tripoli.gui.dataViews.plots.AbstractDataView;
+import org.cirdles.tripoli.gui.dataViews.plots.TicGeneratorForAxes;
+import org.cirdles.tripoli.visualizationUtilities.linePlots.MultiLinePlotBuilder;
+
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+import java.util.ArrayList;
+import java.util.List;
 
 
-public class BasicScatterAndLinePlot extends AbstractDataView {
+public class MultiLinePlotLogX extends AbstractDataView {
 
-    private final ComboPlotBuilder comboPlotBuilder;
-    private double[] yAxisData2;
+    private final MultiLinePlotBuilder multiLinePlotBuilder;
+    private double[][] yData;
 
     /**
      * @param bounds
-     * @param comboPlotBuilder
+     * @param multiLinePlotBuilder
      */
-    public BasicScatterAndLinePlot(Rectangle bounds, ComboPlotBuilder comboPlotBuilder) {
+    public MultiLinePlotLogX(Rectangle bounds, MultiLinePlotBuilder multiLinePlotBuilder) {
         super(bounds, 50, 5);
-        this.comboPlotBuilder = comboPlotBuilder;
+        this.multiLinePlotBuilder = multiLinePlotBuilder;
     }
 
     @Override
     public void preparePanel() {
-        xAxisData = comboPlotBuilder.getxData();
-        yAxisData = comboPlotBuilder.getyData();
-        yAxisData2 = comboPlotBuilder.getyData2();
+        xAxisData = multiLinePlotBuilder.getxData();
+        yData = multiLinePlotBuilder.getyData();
 
-        minX = xAxisData[0];
-        maxX = xAxisData[xAxisData.length - 1];
+        minX = Math.log(xAxisData[0]);
+        maxX = Math.log(xAxisData[xAxisData.length - 1]);
 
-        ticsX = TicGeneratorForAxes.generateTics(minX, maxX, (int) (graphWidth / 40.0));
+        // logarithmic ticsX
+        List<Double> xTicsList = new ArrayList<>();
+        int limitLog = (int) xAxisData[xAxisData.length - 1];
+        for (int logIndex = 1; logIndex <= limitLog; logIndex = logIndex * 10) {
+            xTicsList.add(Math.log(logIndex));
+        }
+        ticsX = new BigDecimal[xTicsList.size()];
+        for (int i = 0; i < xTicsList.size(); i++) {
+            ticsX[i] = new BigDecimal(Double.toString(xTicsList.get(i)));
+        }
         double xMarginStretch = TicGeneratorForAxes.generateMarginAdjustment(minX, maxX, 0.01);
         minX -= xMarginStretch;
         maxX += xMarginStretch;
@@ -39,13 +55,10 @@ public class BasicScatterAndLinePlot extends AbstractDataView {
         minY = Double.MAX_VALUE;
         maxY = -Double.MAX_VALUE;
 
-        for (int i = 0; i < yAxisData.length; i++) {
-            minY = StrictMath.min(minY, yAxisData[i]);
-            maxY = StrictMath.max(maxY, yAxisData[i]);
-            minY = StrictMath.min(minY, yAxisData2[i]);
-            maxY = StrictMath.max(maxY, yAxisData2[i]);
-            if (comboPlotBuilder.isyData2OneSigma()) {
-                minY = StrictMath.min(minY, -yAxisData2[i]);
+        for (double[] yDatum : yData) {
+            for (double v : yDatum) {
+                minY = StrictMath.min(minY, v);
+                maxY = StrictMath.max(maxY, v);
             }
         }
         ticsY = TicGeneratorForAxes.generateTics(minY, maxY, (int) (graphHeight / 15.0));
@@ -73,35 +86,32 @@ public class BasicScatterAndLinePlot extends AbstractDataView {
         text.setFont(Font.font("SansSerif", 12));
         int textWidth = 0;
 
-        showTitle(comboPlotBuilder.getTitle());
+        labelXAxis("Log of Saved Iteration Count");
+        showTitle(multiLinePlotBuilder.getTitle());
 
-        // scatter plot
-        g2d.setLineWidth(0.75);
-        g2d.setStroke(Paint.valueOf("Black"));
-        for (int i = 0; i < xAxisData.length; i++) {
-            g2d.strokeOval(mapX(xAxisData[i]) - 2f, mapY(yAxisData[i]) - 2f, 4f, 4f);
-        }
+        g2d.setLineWidth(2.0);
+        // new line plots
+        // specify cool heatmap
+        double rgbStartRed = 151.0 / 255.0;
+        double rgbStartGreen = 248.0 / 255.0;
+        double rgbStartBlue = 253.0 / 255.0;
+        double rgbEndRed = 231.0 / 255.0;
+        double rgbEndGreen = 56.0 / 255.0;
+        double rgbEndBlue = 244.0 / 255.0;
+        double redDelta = (rgbEndRed - rgbStartRed) / yData.length;
+        double greenDelta = (rgbEndGreen - rgbStartGreen) / yData.length;
+        double blueDelta = (rgbEndBlue - rgbStartBlue) / yData.length;
 
-        // new line plot from yAxisData2
-        g2d.setStroke(Paint.valueOf("red"));
-        g2d.beginPath();
-        g2d.moveTo(mapX(xAxisData[0]), mapY(yAxisData2[0]));
-        for (int i = 0; i < xAxisData.length; i++) {
-            // line tracing through points
-            g2d.lineTo(mapX(xAxisData[i]), mapY(yAxisData2[i]));
-        }
-        g2d.stroke();
-
-        if (comboPlotBuilder.isyData2OneSigma()) {
+        for (int y = 0; y < yData.length; y++) {
+            g2d.setStroke(Color.color(rgbStartRed + redDelta * y, rgbStartGreen + greenDelta * y, rgbStartBlue + blueDelta * y));
             g2d.beginPath();
-            g2d.moveTo(mapX(xAxisData[0]), mapY(-yAxisData2[0]));
+            g2d.moveTo(mapX(Math.log(xAxisData[0])), mapY(yData[y][0]));
             for (int i = 0; i < xAxisData.length; i++) {
                 // line tracing through points
-                g2d.lineTo(mapX(xAxisData[i]), mapY(-yAxisData2[i]));
+                g2d.lineTo(mapX(Math.log(xAxisData[i])), mapY(yData[y][i]));
             }
             g2d.stroke();
         }
-
 
         if (ticsY.length > 1) {
             // border and fill
@@ -119,7 +129,7 @@ public class BasicScatterAndLinePlot extends AbstractDataView {
             float verticalTextShift = 3.2f;
             g2d.setFont(Font.font("SansSerif", 10));
             if (ticsY != null) {
-                for (java.math.BigDecimal bigDecimal : ticsY) {
+                for (BigDecimal bigDecimal : ticsY) {
                     g2d.strokeLine(
                             mapX(minX), mapY(bigDecimal.doubleValue()), mapX(maxX), mapY(bigDecimal.doubleValue()));
 
@@ -127,24 +137,24 @@ public class BasicScatterAndLinePlot extends AbstractDataView {
                     text.setText(bigDecimal.toString());
                     textWidth = (int) text.getLayoutBounds().getWidth();
                     g2d.fillText(text.getText(),//
-                            (float) mapX(minX) - textWidth + 5f,
+                            (float) mapX(minX) - textWidth - 5f,
                             (float) mapY(bigDecimal.doubleValue()) + verticalTextShift);
 
                 }
                 // ticsX
                 if (ticsX != null) {
-                    for (int i = 0; i < ticsX.length - 1; i++) {
+                    for (BigDecimal bigDecimal : ticsX) {
                         try {
                             g2d.strokeLine(
-                                    mapX(ticsX[i].doubleValue()),
+                                    mapX(bigDecimal.doubleValue()),
                                     mapY(ticsY[0].doubleValue()),
-                                    mapX(ticsX[i].doubleValue()),
+                                    mapX(bigDecimal.doubleValue()),
                                     mapY(ticsY[0].doubleValue()) + 5);
 
                             // bottom
-                            String xText = ticsX[i].toPlainString();
+                            String xText = (new BigDecimal(Double.toString(Math.exp(bigDecimal.doubleValue())))).setScale(-1, RoundingMode.HALF_UP).toPlainString();
                             g2d.fillText(xText,
-                                    (float) mapX(ticsX[i].doubleValue()) - 5f,
+                                    (float) mapX(bigDecimal.doubleValue()) - 5f,
                                     (float) mapY(ticsY[0].doubleValue()) + 15);
 
                         } catch (Exception ignored) {
