+++ TripoliApp/src/main/java/org/cirdles/tripoli/gui/dialogs/TripoliMessageDialog.java
@@ -45,7 +45,7 @@ public class TripoliMessageDialog extends Alert {
         initStyle(StageStyle.DECORATED);
         int countOfNewLines = 1;
         for (int i = 0; i < message.length(); i++) {
-            countOfNewLines = countOfNewLines + ((Character.compare(message.charAt(i), '\n') == 0) ? 1 : 0);
+            countOfNewLines = countOfNewLines + ((message.charAt(i) == '\n') ? 1 : 0);
         }
         getDialogPane().setPrefSize(750, 150 + countOfNewLines * 20);
         getDialogPane().setStyle(getDialogPane().getStyle() + ";-fx-font-family: SansSerif Bold;-fx-font-size: 15");
