+++ TripoliCore/src/main/java/org/cirdles/tripoli/sessions/analysis/massSpectrometerModels/dataOutputModels/mcmc/MassSpecOutputDataRecord.java
@@ -66,6 +66,7 @@ public record MassSpecOutputDataRecord(
         double[] baseLineFlagsForRawDataColumn,
         double[] ionCounterFlagsForRawDataColumn,
         MatrixStore<Double>[] allBlockInterpolations,
+        int[][] indicesOfKnotsByBlock,
         int faradayCount,
         int isotopeCount,
         int blockCount,
