+++ TripoliCore/src/main/java/org/cirdles/tripoli/species/SpeciesRecordInterface.java
@@ -23,5 +23,5 @@ public interface SpeciesRecordInterface {
 
     double getNaturalAbundancePercent();
 
-    public String prettyPrintShortForm();
+    String prettyPrintShortForm();
 }
