+++ TripoliCore/src/main/java/org/cirdles/tripoli/utilities/callbacks/LoggingCallbackInterface.java
@@ -1,5 +1,5 @@
 package org.cirdles.tripoli.utilities.callbacks;
 
 public interface LoggingCallbackInterface {
-    public void receiveLoggingSnippet(String loggingSnippet);
+    void receiveLoggingSnippet(String loggingSnippet);
 }
